+++ src/components/v5/common/TableFiltering/TableFiltering.tsx
@@ -1,6 +1,6 @@
 import React, { FC, PropsWithChildren } from 'react';
-
 import { useIntl } from 'react-intl';
+
 import { TableFilteringProps } from './types';
 import Icon from '~shared/Icon/Icon';
